getting closer on interfaces
